+++ conda/core/linked_data.py
@@ -6,6 +6,7 @@ from logging import getLogger
 from os import listdir
 from os.path import isdir, isfile, join
 
+from ..exceptions import CondaDependencyError
 from ..base.constants import UNKNOWN_CHANNEL
 from ..common.compat import itervalues, odict
 from ..gateways.disk.delete import rm_rf
@@ -144,6 +145,6 @@ def get_python_version_for_prefix(prefix):
         return None
     next_record = next(py_record_iter, None)
     if next_record is not None:
-        raise RuntimeError("multiple python records found in prefix %s" % prefix)
+        raise CondaDependencyError("multiple python records found in prefix %s" % prefix)
     else:
         return record.version[:3]